Add EmojiButtonsController to SimpleKeyboardIME
