+++ app/src/main/java/be/scri/fragments/LanguageSettingsFragment.kt
@@ -10,7 +10,6 @@ import android.widget.Button
 import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.activity.addCallback
-import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.viewpager2.widget.ViewPager2
@@ -18,13 +17,11 @@ import be.scri.R
 import be.scri.activities.MainActivity
 import be.scri.databinding.FragmentLanguageSettingsBinding
 import be.scri.helpers.CustomAdapter
-import be.scri.helpers.EmojiButtonController
 import be.scri.models.SwitchItem
 
 class LanguageSettingsFragment : Fragment() {
     private var _binding: FragmentLanguageSettingsBinding? = null
     val binding get() = _binding!!
-    private lateinit var emojiButtonController: EmojiButtonController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -99,8 +96,6 @@ class LanguageSettingsFragment : Fragment() {
                 .commit()
         }
         setupRecyclerView(language)
-        emojiButtonController = EmojiButtonController(requireContext())
-        emojiButtonController.initializeViews(requireActivity() as AppCompatActivity)
     }
 
     private fun setupRecyclerView(language: String) {
@@ -222,7 +217,7 @@ class LanguageSettingsFragment : Fragment() {
         val editor = sharedPref.edit()
         editor.putBoolean("emoji_suggestions_$language", true)
         editor.apply()
-        emojiButtonController.updateButtonVisibility(true)
+        (requireActivity() as MainActivity).getSimpleKeyboardIME()?.updateButtonVisibility(true)
         Toast.makeText(requireContext(), "$language Emoji Autosuggestions on", Toast.LENGTH_SHORT).show()
     }
 
@@ -231,7 +226,7 @@ class LanguageSettingsFragment : Fragment() {
         val editor = sharedPref.edit()
         editor.putBoolean("emoji_suggestions_$language", false)
         editor.apply()
-        emojiButtonController.updateButtonVisibility(false)
+        (requireActivity() as MainActivity).getSimpleKeyboardIME()?.updateButtonVisibility(false)
         Toast.makeText(requireContext(), "$language Emoji Autosuggestions off", Toast.LENGTH_SHORT).show()
     }
